feat(HOF.py): add tracking of unique users per message using a dictionary
This commit adds a dictionary to track the unique users that have reacted to a message. The dictionary is keyed by the message ID and the value is a set of user IDs. The dictionary is updated in the `check_unique_reactions` method. Additionally, the `on_raw_reaction_remove` method was added to also call `check_unique_reactions` when a reaction is removed.